Add personal pages

+++ app/Controller/ContentCategoriesController.php
@@ -10,12 +10,25 @@ App::uses('AppController', 'Controller');
 class ContentCategoriesController extends AppController {
 
     public $publicActions = array('getList');
-
+    
+    public $paginateConditions = array(
+        'title' => array(
+            'type' => 'LIKE',
+            'field' => 'ContentCategory.name',
+        ),
+        'user' => array(
+            'type' => 'LIKE',
+            'field' => 'User.name',
+        ),
+    );
+    
     public function admin_add() {
         $this->helpers[] = 'TinyMCE.TinyMCE';
         $this->set('title_for_layout', 'افزودن مجموعه مطالب');
         $this->set('parents', $this->ContentCategory->generateTreeList());
         if ($this->request->is('post')) {
+            $this->ContentCategory->create();
+            $this->request->data['ContentCategory']['is_lock'] = false;
             if ($this->ContentCategory->save($this->request->data)) {
                 // Save Level for this item
                 $path = $this->ContentCategory->getPath();
@@ -52,12 +65,13 @@ class ContentCategoriesController extends AppController {
         $count = count($id);
         if ($count == 1) {
             $id = current($id);
+            if(! $this->_canEdit($id)){
+                $this->redirect($this->referer());
+            }
             $this->ContentCategory->id = $id;
 
             if ($this->_haveContent($id) || $this->ContentCategory->childCount($id)) {
                 $this->Session->setFlash(SettingsController::read('Error.Code-15'), 'message', array('type' => 'error'));
-            } elseif ($this->ContentCategory->id == 1) {
-                $this->Session->setFlash(SettingsController::read('Error.Code-18'), 'message', array('type' => 'error'));
             } elseif ($this->ContentCategory->delete()) {
                 $this->Session->setFlash('مجموعه با موفقیت حذف شد.', 'message', array('type' => 'success'));
             } else {
@@ -66,16 +80,18 @@ class ContentCategoriesController extends AppController {
         } elseif ($count > 1) {
             $countAffected = 0;
             foreach ($id as $i) {
-                $this->ContentCategory->id = $i;
-                if ($this->_haveContent($i)) {
-                    continue;
-                }
-                if ($this->ContentCategory->childCount($i)) {
-                    continue;
-                }
-
-                if ($this->ContentCategory->delete()) {
-                    $countAffected++;
+                if($this->_canEdit($i)){
+                    $this->ContentCategory->id = $i;
+                    if ($this->_haveContent($i)) {
+                        continue;
+                    }
+                    if ($this->ContentCategory->childCount($i)) {
+                        continue;
+                    }
+    
+                    if ($this->ContentCategory->delete()) {
+                        $countAffected++;
+                    }
                 }
             }
             $this->Session->setFlash($countAffected . ' مجموعه با موفقیت حذف شد.', 'message', array('type' => 'success'));
@@ -90,7 +106,12 @@ class ContentCategoriesController extends AppController {
         if (!$this->ContentCategory->exists()) {
             throw new NotFoundException(SettingsController::read('Error.Code-14'));
         }
+        if(! $this->_canEdit($id)){
+            $this->redirect(array('action' => 'index'));
+        }
         if ($this->request->is('post') || $this->request->is('put')) {
+            
+            $this->request->data['ContentCategory']['is_lock'] = false;
             if ($this->ContentCategory->save($this->request->data)) {
                 // Save Level for this item
                 $path = $this->ContentCategory->getPath();
@@ -105,9 +126,6 @@ class ContentCategoriesController extends AppController {
                 $this->Session->setFlash(SettingsController::read('Error.Code-13'), 'message', array('type' => 'error'));
             }
         } else {
-            if ($this->ContentCategory->id == 1) {
-                $this->Session->setFlash(SettingsController::read('Error.Code-18'), 'message', array('type' => 'error'));
-            }
             $this->set('parents', $this->ContentCategory->generateTreeList());
             $this->request->data = $this->ContentCategory->read();
         }
@@ -143,9 +161,118 @@ class ContentCategoriesController extends AppController {
      * 
      * @return
      */
-    public function getList() {
-        $categories = $this->ContentCategory->find('all', array('conditions' => array('published' => true), 'contain' => false));
+    public function getList($user_id = null) {
+        $conditions = array('published' => true);
+        if($user_id){
+            $conditions['user_id'] = $user_id;
+        }
+        $categories = $this->ContentCategory->find('all', array(
+            'conditions' => $conditions, 
+            'contain' => false
+        ));
         return $categories;
     }
+    
+    /**
+     * This function used from other controllers
+     * 
+     * @param mixed $id
+     * @return
+     */
+    public function _canEdit($id){
+        $this->ContentCategory->recursive = -1;
+        $category = $this->ContentCategory->read(null, $id);
+        
+        if($category['ContentCategory']['is_lock']){
+            $this->Session->setFlash(SettingsController::read('Error.Code-18'), 'message', array('type' => 'error'));
+            return false;
+        }
+        /**
+         * User is Register and he is not owner of this category
+         */
+        if($this->Auth->user('Role.name') == 'Register' and $category['ContentCategory']['user_id'] != $this->Auth->user('id')){            
+            $this->Session->setFlash('مجموعه موردنظر غیرقابل ویرایش است', 'message', array('type' => 'warning'));
+            return false;
+        }
+        return true;
+    }
+    
+    /**
+     * This actions can access from members
+     * 
+     * 
+     * 
+     */
+     public function index(){
+        $this->set('title_for_layout', 'مدیریت مجموعه مطالب');
+        $this->paginate['ContentCategory']['conditions']['ContentCategory.user_id'] = $this->Auth->user('id');
+        $contentCategories = $this->paginate('ContentCategory');
 
+        for ($i = 0; $i < count($contentCategories); $i++) {
+            $contentCategories[$i]['ContentCategory']['contentCount'] = $this->_haveContent($contentCategories[$i]['ContentCategory']['id']);
+        }
+
+        // add this helper for using FilterHelper in Filter Form
+        $this->helpers[] = 'AdminForm';
+        $this->set(compact('contentCategories'));
+     }
+     public function add(){
+        $this->helpers[] = 'TinyMCE.TinyMCE';
+        $this->set('title_for_layout', 'افزودن مجموعه مطالب');
+        $this->set('parents', $this->ContentCategory->generateTreeList(array('user_id' => $this->Auth->user('id'))));
+        if ($this->request->is('post')) {
+            $this->ContentCategory->create();
+            $this->request->data['ContentCategory']['is_lock'] = false;
+            $this->request->data['ContentCategory']['user_id'] = $this->Auth->user('id');
+            $this->request->data['ContentCategory']['published'] = true;
+            $this->request->data['ContentCategory']['access'] = 1;
+            if ($this->ContentCategory->save($this->request->data)) {
+                // Save Level for this item
+                $path = $this->ContentCategory->getPath();
+                // levels starts with 0
+                $this->ContentCategory->saveField('level', count($path) - 1);
+
+                $this->Session->setFlash('مجموعه با موفقیت ذخیره شد.', 'message', array('type' => 'success'));
+                $this->redirect(array('action' => 'index'));
+            } else {
+                $this->Session->setFlash(SettingsController::read('Error.Code-13'), 'message', array('type' => 'error'));
+            }
+        }
+     }
+     public function edit($id){
+        $this->helpers[] = 'TinyMCE.TinyMCE';
+        $this->set('title_for_layout', 'ویرایش مجموعه مطلب');
+        $this->ContentCategory->id = $id;   
+        if (!$this->ContentCategory->exists()) {
+            throw new NotFoundException(SettingsController::read('Error.Code-14'));
+        }
+        if(! $this->_canEdit($id)){
+            $this->redirect(array('action' => 'index'));
+        }
+        if ($this->request->is('post') || $this->request->is('put')) {
+            $this->request->data['ContentCategory']['is_lock'] = false;
+            $this->request->data['ContentCategory']['user_id'] = $this->Auth->user('id');
+            $this->request->data['ContentCategory']['published'] = true;
+            $this->request->data['ContentCategory']['access'] = 1;
+            if ($this->ContentCategory->save($this->request->data)) {
+                // Save Level for this item
+                $path = $this->ContentCategory->getPath();
+                // levels starts with 0
+                $this->ContentCategory->saveField('level', count($path) - 1);
+                // Update level of childrens
+                $this->ContentCategory->updateChildrenLevel();
+
+                $this->Session->setFlash('مجموعه با موفقیت ویرایش شد', 'message', array('type' => 'success'));
+                $this->redirect(array('action' => 'index'));
+            } else {
+                $this->Session->setFlash(SettingsController::read('Error.Code-13'), 'message', array('type' => 'error'));
+            }
+        } else {
+            $this->set('parents', $this->ContentCategory->generateTreeList(array('user_id' => $this->Auth->user('id'))));
+            $this->request->data = $this->ContentCategory->read();
+        }
+     }
+     public function delete(){
+        $this->admin_delete();
+     }
 }
